There are four changes here, mostly to demonstrated features of Golang.
The first change was to move the random number seeding out of the loop.
Seeding only needs to be done once per program run; and repeatedly seeding
from a pseudo random source reduces the randomness of the generator.

The second change was to use range for the for loop. This is a common Go
idiom, that should be adopted "because".

The third was to change the statement type to directly reference the set
(slice) of words that it could substitute rather than an encoding (Int)
which needed to be interpreted programatically.

The fourth was to demonstrate the utility of [3] by adding a new dictionary,
adverb, and using it in a statement.

Signed-off-by: Steve McPolin <smcpolin@vmware.com>